refactor: standardize stores and job actions
