+++ frontend/src/Pages/Authentication/ClientOnboarding/ClientOnboarding.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import useUserStore from '../../../store/userStore.store'
+import useUserStore from '../../../store/useUserStore'
 import Sidebar from './components/Sidebar'
 import MobileHeader from './components/MobileHeader'
 import Step1 from './components/Step1'
