+++ frontend/src/Pages/Dashboard/ClientDashoard/HomePage/Components/QuickActions.jsx
@@ -4,7 +4,7 @@ import {
   ListBulletIcon,
   CreditCardIcon,
 } from '@heroicons/react/24/outline'
-import useClientStore from '../../../../../store/clientStore.store'
+import useClientStore from '../../../../../store/useClientStore'
 const QuickActions = () => {
   const setMainTab = useClientStore(state => state.setMainTab)
   const setJobTab = useClientStore(state => state.setJobTab)
