+++ frontend/src/Pages/Dashboard/ProfessionalDashboard/HomePage/Components/QuickActions.jsx
@@ -5,7 +5,7 @@ import {
   CreditCardIcon,
 } from '@heroicons/react/24/outline'
 
-import useProStore from '../../../../../store/proStore.store'
+import useProStore from '../../../../../store/useProStore'
 
 const QuickActions = () => {
   const setMainTab = useProStore(state => state.setMainTab)
